Split carto theme into sections (#495)
