+++ packages/react-ui/src/theme/sections/typography.js
@@ -1,0 +1,125 @@
+export const themeTypography = {
+  htmlFontSize: 16,
+  fontFamily: 'Montserrat, sans-serif',
+  fontSize: 16,
+  lineHeight: 1.5,
+  fontWeightLight: 300,
+  fontWeightRegular: 400,
+  fontWeightMedium: 500,
+  fontWeightBold: 600,
+  fontSmoothing: 'antialiased',
+  h1: {
+    fontFamily: 'Montserrat, sans-serif',
+    fontWeight: 600,
+    fontSize: '6rem',
+    lineHeight: 1.084,
+    letterSpacing: '-0.016em',
+    fontSmoothing: 'antialiased'
+  },
+  h2: {
+    fontFamily: 'Montserrat, sans-serif',
+    fontWeight: 600,
+    fontSize: '4rem',
+    lineHeight: 1.125,
+    letterSpacing: '-0.008em',
+    fontSmoothing: 'antialiased'
+  },
+  h3: {
+    fontFamily: 'Montserrat, sans-serif',
+    fontWeight: 600,
+    fontSize: '3rem',
+    lineHeight: 1.167,
+    letterSpacing: '0em',
+    fontSmoothing: 'antialiased'
+  },
+  h4: {
+    fontFamily: 'Montserrat, sans-serif',
+    fontWeight: 600,
+    fontSize: '2.125rem',
+    lineHeight: 1.177,
+    letterSpacing: '0.007em',
+    fontSmoothing: 'antialiased'
+  },
+  h5: {
+    fontFamily: 'Montserrat, sans-serif',
+    fontWeight: 600,
+    fontSize: '1.5rem',
+    lineHeight: 1.334,
+    letterSpacing: '0em',
+    fontSmoothing: 'antialiased'
+  },
+  h6: {
+    fontFamily: 'Montserrat, sans-serif',
+    fontWeight: 600,
+    fontSize: '1.25rem',
+    lineHeight: 1.2,
+    letterSpacing: '0.007em',
+    fontSmoothing: 'antialiased'
+  },
+  subtitle1: {
+    fontFamily: 'Montserrat, sans-serif',
+    fontWeight: 600,
+    fontSize: '1rem',
+    lineHeight: 1.5,
+    letterSpacing: '0.009em',
+    fontSmoothing: 'antialiased'
+  },
+  subtitle2: {
+    fontFamily: '"Open Sans", sans-serif',
+    fontWeight: 600,
+    fontSize: '0.875rem',
+    lineHeight: 1.715,
+    letterSpacing: '0.007em',
+    fontSmoothing: 'antialiased'
+  },
+  body1: {
+    fontFamily: '"Open Sans", sans-serif',
+    fontWeight: 400,
+    fontSize: '1rem',
+    lineHeight: 1.5,
+    letterSpacing: '0.028em',
+    fontSmoothing: 'antialiased'
+  },
+  body2: {
+    fontFamily: '"Open Sans", sans-serif',
+    fontWeight: 400,
+    fontSize: '0.875rem',
+    lineHeight: 1.429,
+    letterSpacing: '0.018em',
+    fontSmoothing: 'antialiased'
+  },
+  button: {
+    fontFamily: 'Montserrat, sans-serif',
+    fontWeight: 600,
+    fontSize: '0.875rem',
+    lineHeight: 1.715,
+    letterSpacing: '0.018em',
+    textTransform: 'inherit',
+    fontSmoothing: 'antialiased'
+  },
+  caption: {
+    fontFamily: '"Open Sans", sans-serif',
+    fontWeight: 600,
+    fontSize: '0.75rem',
+    lineHeight: 1.334,
+    letterSpacing: '0.017em',
+    fontSmoothing: 'antialiased'
+  },
+  overline: {
+    fontFamily: '"Open Sans", sans-serif',
+    fontWeight: 400,
+    fontSize: '0.625rem',
+    lineHeight: 1.6,
+    letterSpacing: '0.150em',
+    textTransform: 'uppercase',
+    fontSmoothing: 'antialiased'
+  },
+  charts: {
+    fontFamily: '"Open Sans", sans-serif',
+    fontWeight: 400,
+    fontSize: 10,
+    lineHeight: 1.6,
+    letterSpacing: '0.150em',
+    fontSmoothing: 'antialiased'
+  }
+};
