+++ packages/react-ui/src/theme/themeConstants.js
@@ -1,0 +1,2 @@
+// Common
+export const SPACING = 8;
