+++ packages/react-ui/src/theme/themeUtils.js
@@ -1,0 +1,4 @@
+import createSpacing from '@material-ui/core/styles/createSpacing';
+import { SPACING } from './themeConstants';
+
+export const spacing = createSpacing(SPACING);
